pyproject.toml
[project]
name = "smartnote"
version = "0.1.0"
description = ""
authors = [
    {name = "leeseungjun",email = "sjun.leee@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "typer[all] (>=0.21.1,<0.22.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "anthropic (>=0.79.0,<0.80.0)",
    "langchain (>=1.2.9,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)",
    "langchain-anthropic (>=1.3.2,<2.0.0)",
    "chromadb (>=1.5.0,<2.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "markdown2 (>=2.5.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "smartnote", from = "src"}]

[tool.poetry.scripts]
smartnote = "smartnote.cli:app" 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
